pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobhunter-ai"
version = "0.1.0"
description = "AI-powered job hunting assistant - search, match, and apply smarter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Edwin Isac", email = "edwinisac007@gmail.com" }
]
keywords = ["job-search", "ai", "career", "automation", "job-hunting", "cover-letter", "resume"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pyyaml>=6.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
]

[project.scripts]
jobhunter = "jobhunter.cli:main"

[project.urls]
Homepage = "https://github.com/edwiniac/jobhunter-ai"
Repository = "https://github.com/edwiniac/jobhunter-ai"
Issues = "https://github.com/edwiniac/jobhunter-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["jobhunter"]
